+++ www/src/components/page-header/index.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
 import { Title, Text } from '@thumbtack/thumbprint-react';
 import classNames from 'classnames';